// src/enums/EggGroupName.ts
export enum EggGroupName {
  Monster = "Monster Group",
  Water1 = "Water 1 Group",
  Bug = "Bug Group",
  Flying = "Flying Group",
  Field = "Field Group",
  Fairy = "Fairy Group",
  Grass = "Grass Group",
  HumanLike = "Human-Like Group",
  Water3 = "Water 3 Group",
  Mineral = "Mineral Group",
  Amorphous = "Amorphous Group",
  Water2 = "Water 2 Group",
  Ditto = "Ditto Group",
  Dragon = "Dragon Group",
  NoEggs = "No Eggs Discovered Group"
}